02_src/gcp/cloud_functions/01_gcs_to_bq/requirements.txt
pandas
pandas-gbq
gcsfs
fsspec
google-cloud-bigquery
google-cloud-storage